Refactored the domain tests to be in the proper package
