+++ src/test/java/com/sethhaskellcondie/thegamepensiveapi/domain/entity/system/SystemServiceTests.java
@@ -1,11 +1,7 @@
-package com.sethhaskellcondie.thegamepensiveapi.domain.system;
+package com.sethhaskellcondie.thegamepensiveapi.domain.entity.system;
 
 import com.sethhaskellcondie.thegamepensiveapi.domain.entity.EntityRepository;
 import com.sethhaskellcondie.thegamepensiveapi.domain.entity.EntityServiceTests;
-import com.sethhaskellcondie.thegamepensiveapi.domain.entity.system.System;
-import com.sethhaskellcondie.thegamepensiveapi.domain.entity.system.SystemRequestDto;
-import com.sethhaskellcondie.thegamepensiveapi.domain.entity.system.SystemResponseDto;
-import com.sethhaskellcondie.thegamepensiveapi.domain.entity.system.SystemService;
 import com.sethhaskellcondie.thegamepensiveapi.domain.filter.FilterService;
 
 import static org.mockito.Mockito.mock;
